Clean up og refactor
