+++ src/Banner/Banner.tsx
@@ -3,9 +3,8 @@ import './Banner.less';
 import { RouteComponentProps, withRouter } from 'react-router-dom';
 import Bedriftsmeny from '@navikt/bedriftsmeny';
 import '@navikt/bedriftsmeny/lib/bedriftsmeny.css';
-import { AltinnOrganisasjon } from '../api/organisasjonstre/organisasjonstre-utils';
 import { RestStatus } from '../api/api-utils';
-import { RestAltinnOrganisasjoner } from '../api/organisasjonstre/organisasjonstre-api';
+import { AltinnOrganisasjon, RestAltinnOrganisasjoner } from '../api/altinnorganisasjon/altinnorganisasjon-api';
 
 interface Props {
     tittel: string;
