+++ src/mocking/mock.ts
@@ -4,7 +4,7 @@ import { getOrganisasjonerMock } from './organisasjoner';
 import { getSykefraværshistorikkMock } from './sykefraværshistorikk';
 
 const MOCK_MIN_SIDE_ARBEIDSGIVER = true;
-const MOCK_SYKEFRAVÆRSSTATISTIKK_API = false;
+const MOCK_SYKEFRAVÆRSSTATISTIKK_API = true;
 const MOCK_ENHETSREGISTERET = true;
 const MOCK_FEATURE_TOGGLES = true;
