+++ src/mocking/organisasjoner.ts
@@ -1,5 +1,5 @@
 import { organisasjonstreMock } from './organisasjonstre';
-import { AltinnOrganisasjon } from '../api/organisasjonstre/organisasjonstre-utils';
+import { AltinnOrganisasjon } from '../api/altinnorganisasjon/altinnorganisasjon-api';
 
 export const getOrganisasjonerMock = (): AltinnOrganisasjon[] => {
     return organisasjonstreMock;
